requirements.txt
APScheduler==3.6.3
astor==0.8.1
astpretty==3.0.0
atomicwrites==1.4.1
attrs==22.1.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
clarifai-grpc==8.11.0
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
emoji==2.2.0
eradicate==2.1.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.51.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mr-proper==0.0.7
opencv-python==4.6.0.66
packaging==21.3
pluggy==1.0.0
protobuf==3.19.3
psycopg-binary==3.1.6
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.1.1
pyflakes==2.5.0
pymongo==4.3.3
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.8.0
pytest==7.0.1
pytest-sugar==0.9.4
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.44
stdlib-list==0.8.0
termcolor==2.1.1
tomli==2.0.1
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
alembic==1.9.1
APScheduler==3.6.3
astor==0.8.1
astpretty==3.0.0
atomicwrites==1.4.1
attrs==22.1.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
clarifai-grpc==8.11.0
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
emoji==2.2.0
ephem==4.1.3
eradicate==2.1.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.51.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mr-proper==0.0.7
mypy==0.991
mypy-extensions==0.4.3
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
pluggy==1.0.0
protobuf==3.19.3
psycopg-binary==3.1.6
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.1.1
pyflakes==2.5.0
pymongo==4.3.3
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.8.0
pytest==7.0.1
pytest-sugar==0.9.4
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.44
stdlib-list==0.8.0
termcolor==2.1.1
tomli==2.0.1
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1